Setting up JSDoc
