+++ index.html
@@ -26,10 +26,10 @@
           </div>
           <div class="col">
             <label for="password">Password:</label>
-            <input name="password" type="password" class="input mt-1 mb-4" required>
+            <input name="password" type="password" class="input mt-1" required>
           </div>
-          <div id="errorMessage" class="error-message mb-2" style="display:none;"></div>
-          <button class="btn-primary mb-2">Login</button>
+          <div id="errorMessage" class="error-message mt-2" style="display:none;"></div>
+          <button class="btn-primary mt-4 mb-2">Login</button>
         </form>
         <button class="btn-primary-outlined" onclick="location.href='register/index.html'">Register</button>         
       </div>
