+++ js/home-feed.mjs
@@ -10,6 +10,15 @@ import { displayErrorMessage } from "./utils.mjs/displayError.mjs";
 
 let postsArray = [];
 
+/**
+ * Refreshes the feed by loading posts if the user is logged in.
+ * Displays login screen otherwise.
+ * @async
+ * @function
+ * @returns {Promise<void>} A promise that resolves when the feed is refreshed.
+ * @throws {Error} Throws an error if the request fails or if an error occurs in displaying posts.
+ */
+
 export async function refreshFeed() {
   try {
     const isLoggedIn = checkUserLoggedIn();
@@ -36,6 +45,11 @@ if (mySelectElement) {
   });
 };
 
+/**
+ * Handles the selection change event on sorting dropdown.
+ * @param {HTMLSelectElement} selection The dropdown select element.
+ */
+
 function onSelected(selection) {
   let value = selection.value;
   if (value === 'Newest') {
@@ -45,10 +59,18 @@ function onSelected(selection) {
   }
 }
 
+/**
+ * Sorts the posts array by the newest posts.
+ */
+
 function sortByNewest() {
   postsArray.sort((a, b) => new Date(b.created) - new Date(a.created));
   displayPosts(postsArray);
 }
+
+/**
+ * Sorts the posts array by the oldest posts.
+ */
 
 function sortByOldest() {
   postsArray.sort((a, b) => new Date(a.created) - new Date(b.created));
@@ -62,12 +84,19 @@ if (searchInputElement) {
     searchInputElement.addEventListener('input', debounce(searchPosts, 1000));
 }
 
-function searchPosts() {
-  const searchText = document.getElementById('searchInput').value.toLowerCase();
-  const filteredPosts = postsArray.filter(post => post.title.toLowerCase().includes(searchText));
-  displayPosts(filteredPosts);
-}
+/**
+ * Filters posts based on the search input.
+ */
 
+function searchPosts() {
+  try {
+    const searchText = document.getElementById('searchInput').value.toLowerCase();
+    const filteredPosts = postsArray.filter(post => post.title.toLowerCase().includes(searchText));
+    displayPosts(filteredPosts);
+  } catch (error) {
+    displayErrorMessage("Error occurred during search.");
+  }
+}
 
 attachPostSubmitListener();
 refreshFeed();
