+++ js/utils.mjs/debounce.mjs
@@ -1,3 +1,12 @@
+/**
+ * Creates a debounced function that delays invoking `func` until after `delay` milliseconds have elapsed
+ * since the last time the debounced function was invoked.
+ * 
+ * @param {Function} func - The function to debounce.
+ * @param {number} delay - The number of milliseconds to delay.
+ * @returns {Function} A new debounced function.
+ */
+
 function debounce(func, delay) {
   let debounceTimer;
   return function() {
