+++ register/index.html
@@ -34,6 +34,7 @@
             <label for="password">Password:</label>
             <input name="password" type="password" class="input mt-1 mb-4" required minlength="8">
           </div>
+          <div id="errorMessage" class="error-message mt-2" style="display:none;"></div>
           <button class="btn-primary mb-2">Create</button>
         </form>          
         <button class="btn-primary-outlined" onclick="location.href='../index.html'">Back</button>
