+++ src/scss/_utilities.scss
@@ -102,6 +102,10 @@
   margin-top: 20px;
 }
 
+.mt-4 {
+  margin-top: 30px;
+}
+
 // Margin-bottom
 
 .mb-1 {
