+++ src/scss/components/_card.scss
@@ -5,7 +5,6 @@
   box-shadow: 0px 0px 10px -2px rgba(0, 0, 0, 0.25);
   padding: 5px 0px 5px 0px;
   border-radius: 5px;
-
     h2 {
       font-size: 14px;
     }
